more pokedex file splitting
