+++ src/ui/pages/pokedex/PokedexSidebar.css
@@ -1,0 +1,85 @@
+.pokedex-sidebar {
+  width: var(--sidebar-width);
+  padding: 0 calc(2 * var(--padding-radius-sm-lg));
+  box-sizing: border-box;
+  height: 100%;
+  background-color: var(--color-panel);
+  box-shadow:
+    rgba(50, 50, 93, 0.25) 0px 30px 60px -12px inset,
+    rgba(0, 0, 0, 0.3) 0px 18px 36px -18px inset;
+  overflow-y: auto;
+  button {
+    box-shadow:
+      rgba(0, 0, 0, 0.17) 0px -23px 25px 0px inset,
+      rgba(0, 0, 0, 0.15) 0px -36px 30px 0px inset,
+      rgba(0, 0, 0, 0.1) 0px -79px 40px 0px inset,
+      rgba(0, 0, 0, 0.06) 0px 2px 1px,
+      rgba(0, 0, 0, 0.09) 0px 4px 2px,
+      rgba(0, 0, 0, 0.09) 0px 8px 4px,
+      rgba(0, 0, 0, 0.09) 0px 16px 8px,
+      rgba(0, 0, 0, 0.09) 0px 32px 16px;
+  }
+}
+
+.pokedex-sidebar-inner {
+  width: 100%;
+  position: relative;
+}
+
+.pokedex-sidebar-button {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: calc(var(--sidebar-width) - 2.5rem);
+  margin-left: 0.5rem;
+  height: 2.25rem !important;
+  display: flex;
+  flex-direction: row;
+  justify-content: start;
+  align-items: center;
+  border: none;
+  border-radius: 2rem;
+  gap: var(--padding-radius-sm-lg);
+  --button-box-shadow:
+    rgba(0, 0, 0, 0.17) 0px -23px 25px 0px inset, rgba(0, 0, 0, 0.15) 0px -36px 30px 0px inset,
+    rgba(0, 0, 0, 0.1) 0px -79px 40px 0px inset, rgba(0, 0, 0, 0.06) 0px 2px 1px,
+    rgba(0, 0, 0, 0.09) 0px 4px 2px, rgba(0, 0, 0, 0.09) 0px 8px 4px,
+    rgba(0, 0, 0, 0.09) 0px 16px 8px, rgba(0, 0, 0, 0.09) 0px 32px 16px;
+  background-color: var(--gray-7);
+}
+
+.pokedex-sidebar-button-selected {
+  background-color: var(--accent-10);
+  color: white;
+  font-weight: bold;
+}
+
+.pokedex-sidebar-button-caught {
+  background-color: var(--accent-6);
+  font-weight: bold;
+}
+
+.pokedex-sidebar-shiny-icon {
+  height: 1.5rem;
+  margin-right: var(--padding-radius-xs-sm);
+}
+
+.pokedex-icon-container {
+  width: 3rem;
+  height: 2.5rem;
+  padding: 0 0.5rem;
+  background-color: var(--gray-7);
+  border-radius: 2rem;
+  box-shadow:
+    0px 1px 1px 1px rgb(0 0 0 / 20%),
+    0px 0.0125rem 0.0125rem 0px rgb(0 0 0 / 14%),
+    0px 0.0125rem 0.125rem 0px rgb(0 0 0 / 12%);
+  margin-left: -1rem;
+  overflow: visible;
+  position: relative;
+
+  .pokemon-icon-image {
+    width: 2.5rem;
+    aspect-ratio: 1;
+  }
+}
